% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codontest-package.R
\docType{package}
\name{codontest-package}
\alias{codontest}
\alias{codontest-package}
\title{codontest: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
