% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qux.R
\name{read_npy}
\alias{read_npy}
\title{Read a matrix from a '.npy' file}
\usage{
read_npy(file, do_transpose = TRUE)
}
\arguments{
\item{file}{Path to '.npy' file}

\item{do_transpose}{Logical. Should the loaded matrix be transposed?}
}
\value{
Numeric matrix
}
\description{
Read a matrix from a '.npy' file
}
