% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qux.R
\name{save_npy}
\alias{save_npy}
\title{Save a matrix to a '.npy' file}
\usage{
save_npy(x, file = tempfile(fileext = ".npy"))
}
\arguments{
\item{x}{Numeric matrix}

\item{file}{Path to '.npy' file}
}
\value{
A list containing information about the saved array.
}
\description{
Save a matrix to a '.npy' file
}
